Move go-specific plugins to ftplugin folder
No filetype plugin should try to do anything during vim's startup (the
plugin folder is sourced on start). The test script was moved into a
folder that vim ignores so to not confuse it's purpose. The oracle
script was moved here as well, as it's not properly setup as a plugin
and should only be sourced when requested.