Fix an error in complete.vim
Prior to this change, placing a cursor over the string `"'"` with completion enabled caused vim to report the following error:

    Error detected while processing function go#complete#Info..go#complete#GetInfo:
    line   29:
    E15: Invalid expression: "\>'

This is because `<cword>` expanded to `"'"`, which caused the filter expression to evaluate to `v:val =~ '\<"'"\>'`. The unescaped single-quote in the middle terminated the string, and left the characters `"\>'` dangling as an invalid expression. This commit fixes that problem by escaping any single quotes in `<cword>`.