Parse and add 'fatal errors' to quickfix list
This fixes two existing problems:

1.) If `ShowErrors` was run with the quickfix list already populated it
would not update the quickfix list if the new output contained a fatal
error. e.g. test is fixed, new run results in fatal error, the now-fixed
test is still shown in the quickfix list.

2.) It shows stacktraces without the context of the `fatal error`.
Before this changed, `ShowErrors` parsed stacktraces (by accident, I
assume) and added them to the quickfix list. But without the `fatal
error:` line, it seemed like these stacktraces belonged to a failing
test.

This change adds parsing of "fatal error" lines to `ShowErrors`. It
checks if a line contains `fatal error:` and if so, it handles the line
like a error and adds it to the quickfix list.

The result is that now fatal errors are handled like regular failing
test: they update the quickfix list.