fix GoDoc if package name differs from import path
when package with "import path" e.g. github.com/bitly/go-simplejson is accessible by default with other name (in this case simplejson),
GoDoc cannot find this because it cannot dereference "package name" to "import path" because of how go#tool#Imports creates mapping between "package_name" and "import path"

this merge tries to infer real "package name" that if GoDoc is placed on  eg. simplejson.NewJson which is located in "github.com/bitly/go-simplejson" - it correctly finds out the correct path

example of non-working code, before path
```
import "github.com/bitly/go-simplejson"

func main(){
    json, err := simplejson.NewJson()
}
```

I am aware that this is problem with libraray that not sticks with convention of giving the same name to package and last folder in import path - but it looks it is quite common.