Use shiftwidth() instead of 'shiftwidth'
I'm using vim-sleuth[1] from tpope to automatically detect indent
settings for files that I edit. The latest version sets 'shiftwidth' to
zero for files that are detected to be indented only with hard tabs, so
that the value of 'tabstop' is used instead of 'shiftwidth'. This
change[2] was done to allow people setting 'tabstop' to their preferred
value, and to have 'shiftwidth' automatically follow.

The vim documentation about 'shiftwidth' states:

   When zero the 'ts' value will be used.  Use the shiftwidth()
   function to get the effective shiftwidth value.

Therefore this commit changes using 'sw' directly and instead calls
shiftwidth().

[1] https://github.com/tpope/vim-sleuth
[2] https://github.com/tpope/vim-sleuth/commit/b6b4c3b237678f1214ea2eca0e3e50eaafb30e3a