fmt: fix preventing other windows to be changed, closes #432
Thanks to @mhinz

Deleting all lines (%delete) from the buffer would Vim cause to change
the same buffer in other views to be changed as well, which would cause
to change the view and cursor position.

Instead we directly replace the content now, instead of deleting and
appending. This doesn't cause the other view to be changed at all,
because we don't delete all files anymore.