commands: use bang attribute to jump to errors
Previously we were using a custom global variable to control the
behavior of jumpting to an error when we called any command in cmd.vim
(such as :GoInstall, :GoTest, etc..)

However we don't need this actually because the default :make command
in vim has the behavior of no jumping to the error when the `!` attribute
is passed to the command, i.e: `:make!`. We have this feature already
for `:GoBuild` and `:GoRun` but those were also conflicting with out
manual `g:go_jump_to_error` setting. Basically even if you call
`:GoBuild` (without `!`) it would not jump if `g:go_jump_to_error` was
set to `0`.

This PR adds the bang attribute, removes the `go_jump_to_error` setting
for the following commands:

* `:GoBuild`
* `:GoRun`
* `:GoInstall`
* `:GoTest`
* `:GoTestFunc`
* `:GoTestCompile`
* `:GoVet`
* `:GoCoverage`

Now basically the behavior is just like `:make`:

* `GoBuild` jumps to the error
* `GoBuild!` does not jump to the error