Drop bad filenames from qflist in GoRun
Fix for https://github.com/fatih/vim-go/issues/287

When a line of output is parsed incorrectly as an error message, the
quickfix feature of Vim creates an unlisted buffer containing a bad
filename. On jumping to this item with `cc`, a blank, unsaved file is
opened in the active window. It is never useful to jump to a nonexistent
file from quickfix and users unused to navigating buffers in Vim may
have trouble figuring out what happened.

Since the line of output that was parsed may still be an important error
message, rather than silently dropping it from the quickfix window, echo
a message for each bad filename dropped in this way. (Known issue: If
`cc 1` jumps to a new buffer immediately, the message `"{filename}"
{linenumber}L, {colnumber}C` hides the drop messages.)

While looping through qflist items, avoid checking the same `filename`
more than once by remembering the return values of previous calls to
`filereadable` during the current execution of `:GoRun`.

Add comments to explain the compiler errorformat strings. Patterns must
be appended in separate statements in order to use inline comments here.