util: add auto quickfix window height for errors [GH-592]
Right now we are just calling `cwindow` without any arguments. By
default Vim opens a view with an height of 10. But this doesn't look
good if you have only one single error. It also takes display estate for
small screens.

This PR introduces a new feature, which automatically resizes the
quickfix error for the number of errors. However to prevent consuming
all the view, it's capped at 10 lines (the default Vim value).

We also have a new setting, `g:go_quickfix_height` which let us to give
an explicit window height. If set, automatic resize is enabled and the
height will be always the value of `g:go_quickfix_height`.

Closes #592